extract_utils: Use vdexExtractor and oatdump for deodexing
oatdump is an AOSP tool that can be used to extract dex files from
oat files, and is maintained by Google upstream, unlike smali.

vdexExtractor is a tool made by anestisb that is written in C++
with code taken directly from art. However, anestisb has also added
a quicken decompiler to oatdump that was merged upstream, so we only
need vdexExtractor for 8.1 and 8.0.

Both of these tools successfully deodex apps and frameworks jars on
8.1, which was the initial reason for this switch.

Change-Id: Ic2cf2dc627a1ad2fa4d500d02d9eac8b8a9577b5
Signed-off-by: Joe Maples <joe@frap129.org>
Signed-off-by: spezi77 <spezi77@gmx.com>